qtnegf module factorized
