+++ NCexamples/qtnegf/Makefile
@@ -1,0 +1,2 @@
+clean:
+	rm -r [1-3]* *fdf *tcl
