+++ NCexamples/qtnegf/input.yaml
