--- NCexamples/qtransport/Makefile
@@ -1,2 +1,0 @@
-clean:
-	rm -r 1elec 2channel 3postprocess *fdf
