+++ simulators/siesta/siesta_default/model/elec/Au_STRUCT_rigt.fdf
