--- solvers/qttransport/translab.py
@@ -1,1 +1,0 @@
-__init__.py
